requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
pypdf2==3.0.1
python-multipart==0.0.6
faiss-cpu==1.7.4
sentence-transformers==2.2.2
rank-bm25==0.2.1
groq==0.3.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
redis==5.0.1
langchain-text-splitters==0.0.1